requirements.txt
asgiref==3.9.1
colorama==0.4.6
Django==5.2.4
django-cleanup==9.0.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
factory_boy==3.3.3
Faker==37.5.3
iniconfig==2.1.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-django==4.11.1
sqlparse==0.5.3
tzdata==2025.2
